refactor styles functions
