+++ src/plugin/pullStyles.ts
@@ -1,137 +1,11 @@
 /* eslint-disable no-param-reassign */
 import {figmaRGBToHex} from '@figma-plugin/helpers';
-import {convertToTokenArray} from '../utils/convertTokens';
 import {ColorToken, TypographyToken} from '../../types/propertyTypes';
 import {slugify} from '../app/components/utils';
-import {
-    convertFigmaGradientToString,
-    convertFigmaToLetterSpacing,
-    convertFigmaToLineHeight,
-    convertLetterSpacingToFigma,
-    convertLineHeightToFigma,
-    convertStringToFigmaGradient,
-    convertToFigmaColor,
-} from './helpers';
+import {convertFigmaGradientToString, convertFigmaToLetterSpacing, convertFigmaToLineHeight} from './helpers';
 import {notifyStyleValues} from './notifiers';
 
-export const setTextValuesOnTarget = async (target, token) => {
-    const {value, description} = token;
-    const {fontFamily, fontWeight, fontSize, lineHeight, letterSpacing, paragraphSpacing} = value;
-    const family = fontFamily || target.fontName.family;
-    const style = fontWeight || target.fontName.style;
-    await figma.loadFontAsync({family, style});
-
-    if (fontFamily && fontWeight) {
-        target.fontName = {
-            family,
-            style,
-        };
-    }
-
-    if (fontSize) {
-        target.fontSize = Number(fontSize);
-    }
-    if (lineHeight) {
-        target.lineHeight = convertLineHeightToFigma(lineHeight);
-    }
-    if (letterSpacing) {
-        target.letterSpacing = convertLetterSpacingToFigma(letterSpacing);
-    }
-    if (paragraphSpacing) {
-        target.paragraphSpacing = Number(paragraphSpacing);
-    }
-    if (description) {
-        target.description = description;
-    }
-};
-
-const setColorValuesOnTarget = (target, token) => {
-    try {
-        const {description, value} = token;
-        if (value.startsWith('linear-gradient')) {
-            const {gradientStops, gradientTransform} = convertStringToFigmaGradient(value);
-            const newPaint = {
-                type: 'GRADIENT_LINEAR',
-                gradientTransform,
-                gradientStops,
-            };
-            target.paints = [newPaint];
-        } else {
-            const {color, opacity} = convertToFigmaColor(value);
-            target.paints = [{color, opacity, type: 'SOLID'}];
-        }
-
-        if (description) {
-            target.description = description;
-        }
-    } catch (e) {
-        console.error('Error setting color', e);
-    }
-};
-
-const updateColorStyles = (colorTokens, shouldCreate = false) => {
-    // Iterate over colorTokens to create objects that match figma styles
-    const colorTokenArray = convertToTokenArray(colorTokens, true);
-    const paints = figma.getLocalPaintStyles();
-
-    colorTokenArray.map(([key, value]: [string, ColorToken]) => {
-        let matchingStyles = [];
-        if (paints.length > 0) {
-            matchingStyles = paints.filter((n) => n.name === key);
-        }
-        if (matchingStyles.length) {
-            setColorValuesOnTarget(matchingStyles[0], value);
-        } else if (shouldCreate) {
-            const style = figma.createPaintStyle();
-            style.name = key;
-            setColorValuesOnTarget(style, value);
-        }
-    });
-};
-
-const updateTextStyles = (textTokens, shouldCreate = false) => {
-    // Iterate over textTokens to create objects that match figma styles
-    const textTokenArray = convertToTokenArray(textTokens, true);
-    const textStyles = figma.getLocalTextStyles();
-
-    textTokenArray.map(([key, value]: [string, TypographyToken]) => {
-        let matchingStyles = [];
-        if (textStyles.length > 0) {
-            matchingStyles = textStyles.filter((n) => {
-                const splitName = n.name.split('/').map((name) => name.trim());
-                const splitKey = key.split('/').map((name) => name.trim());
-
-                if (splitKey[splitKey.length - 1] === 'value') {
-                    splitKey.pop();
-                }
-                const trimmedName = splitName.join('/');
-                const trimmedKey = splitKey.join('/');
-
-                return trimmedName === trimmedKey;
-            });
-        }
-
-        if (matchingStyles.length) {
-            setTextValuesOnTarget(matchingStyles[0], value);
-        } else if (shouldCreate) {
-            const style = figma.createTextStyle();
-            style.name = key;
-            setTextValuesOnTarget(style, value);
-        }
-    });
-};
-
-export function updateStyles(tokens, shouldCreate = false): void {
-    if (!tokens.colors && !tokens.typography) return;
-    if (tokens.colors) {
-        updateColorStyles(tokens.colors, shouldCreate);
-    }
-    if (tokens.typography) {
-        updateTextStyles(tokens.typography, shouldCreate);
-    }
-}
-
-export function pullStyles(styleTypes): void {
+export default function pullStyles(styleTypes): void {
     let colors;
     let typography;
     let fontFamilies;
