+++ src/plugin/updateNode.ts
@@ -1,7 +1,7 @@
 import {convertToFigmaColor} from './helpers';
-import {setTextValuesOnTarget} from './styles';
+import setTextValuesOnTarget from './setTextValuesOnTarget';
 
-export async function setValuesOnNode(node, values, data) {
+export default async function setValuesOnNode(node, values, data) {
     // BORDER RADIUS
     if (values.borderRadius) {
         if (typeof node.cornerRadius !== 'undefined') {
@@ -167,84 +167,5 @@ export async function setValuesOnNode(node, values, data) {
         if (typeof node.itemSpacing !== 'undefined') {
             node.itemSpacing = Number(values.itemSpacing);
         }
-    }
-}
-
-export async function removeValuesFromNode(node, prop) {
-    // BORDER RADIUS
-    switch (prop) {
-        case 'borderRadius':
-            if (typeof node.cornerRadius !== 'undefined') {
-                node.cornerRadius = 0;
-            }
-            break;
-        case 'borderRadiusTopLeft':
-            if (typeof node.cornerRadius !== 'undefined') {
-                node.topLeftRadius = 0;
-            }
-            break;
-        case 'borderRadiusTopRight':
-            if (typeof node.cornerRadius !== 'undefined') {
-                node.topRightRadius = 0;
-            }
-            break;
-        case 'borderRadiusBottomRight':
-            if (typeof node.cornerRadius !== 'undefined') {
-                node.bottomRightRadius = 0;
-            }
-            break;
-        case 'borderRadiusBottomLeft':
-            if (typeof node.cornerRadius !== 'undefined') {
-                node.bottomLeftRadius = 0;
-            }
-            break;
-        case 'borderWidth':
-            if (typeof node.strokeWeight !== 'undefined') {
-                node.strokeWeight = 0;
-            }
-            break;
-        case 'opacity':
-            if (typeof node.opacity !== 'undefined') {
-                node.opacity = 1;
-            }
-            break;
-        case 'fill':
-            if (typeof node.fills !== 'undefined') {
-                node.fills = [];
-            }
-            break;
-        case 'border':
-            if (typeof node.strokes !== 'undefined') {
-                node.strokes = [];
-            }
-            break;
-        case 'spacing':
-            if (typeof node.paddingLeft !== 'undefined') {
-                node.paddingLeft = 0;
-                node.paddingRight = 0;
-                node.paddingTop = 0;
-                node.paddingBottom = 0;
-                node.itemSpacing = 0;
-            }
-            break;
-        case 'horizontalPadding':
-            if (typeof node.paddingLeft !== 'undefined') {
-                node.paddingLeft = 0;
-                node.paddingRight = 0;
-            }
-            break;
-        case 'verticalPadding':
-            if (typeof node.paddingTop !== 'undefined') {
-                node.paddingTop = 0;
-                node.paddingBottom = 0;
-            }
-            break;
-        case 'itemSpacing':
-            if (typeof node.itemSpacing !== 'undefined') {
-                node.itemSpacing = 0;
-            }
-            break;
-        default:
-            break;
     }
 }
