+++ src/plugin/updateStyles.ts
@@ -1,0 +1,12 @@
+import updateColorStyles from './updateColorStyles';
+import updateTextStyles from './updateTextStyles';
+
+export default function updateStyles(tokens, shouldCreate = false): void {
+    if (!tokens.colors && !tokens.typography) return;
+    if (tokens.colors) {
+        updateColorStyles(tokens.colors, shouldCreate);
+    }
+    if (tokens.typography) {
+        updateTextStyles(tokens.typography, shouldCreate);
+    }
+}
