+++ tests/__mocks__/figmaMock.js
@@ -7,4 +7,6 @@ global.figma = {
     ui: {
         postMessage: jest.fn(() => Promise.resolve({})),
     },
+    getLocalPaintStyles: jest.fn(() => Promise.resolve()),
+    getLocalTextStyles: jest.fn(() => Promise.resolve()),
 };
